requirements.txt
asgiref==3.10.0
attrs==25.4.0
black==25.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
Django==5.2.7
djangorestframework==3.16.1
drf-spectacular==0.28.0
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
psycopg2-binary==2.9.11
python-dotenv==1.1.1
pytokens==0.2.0
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.27.1
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
